100daysofpython/twitter-python/requirements.txt
autopep8==1.5.4
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
decorator==4.4.2
future==0.18.2
idna==2.10
oauthlib==3.1.0
pycodestyle==2.6.0
PySocks==1.7.1
python-dotenv==0.15.0
ratelim==0.1.6
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
toml==0.10.2
tweepy==3.9.0
urllib3==1.26.2
